Refactor TopNav
converting it into a function component and extract CorporaDropdown into
separate file
